+++ src/ios/Protocols/OSPMTHandlerDelegate.swift
@@ -1,0 +1,5 @@
+typealias OSPMTCompletionHandler = (Result<Bool, OSPMTError>) -> Void
+
+protocol OSPMTHandlerDelegate {
+    func setupConfiguration() -> Result<String, OSPMTError>
+}